feat: adicionar regras de estilo e melhorar testes de expiração de cache
