+++ src/Routing/StaticRouteManager.php
@@ -227,100 +227,3 @@ class StaticRouteManager
         // Esta é a vantagem da abordagem explícita
     }
 }
-
-/**
- * Mock Request para capturar responses estáticas
- */
-class MockRequest
-{
-    /**
-     * @param mixed $default
-     * @return mixed
-     */
-    public function param(string $name, $default = null)
-    {
-        return $default;
-    }
-
-    /**
-     * @param mixed $default
-     * @return mixed
-     */
-    public function get(string $name, $default = null)
-    {
-        return $default;
-    }
-
-    /**
-     * @param array<mixed> $args
-     * @return mixed
-     */
-    public function __call(string $method, array $args)
-    {
-        return null;
-    }
-}
-
-/**
- * Mock Response para capturar responses estáticas
- */
-class MockResponse
-{
-    private string $content = '';
-    /** @var array<string, string> */
-    private array $headers = [];
-    private int $statusCode = 200;
-
-    /**
-     * @param array<mixed> $data
-     */
-    public function json(array $data): self
-    {
-        $json = json_encode($data);
-        $this->content = $json !== false ? $json : '';
-        $this->headers['Content-Type'] = 'application/json';
-        return $this;
-    }
-
-    public function send(string $content): self
-    {
-        $this->content = $content;
-        return $this;
-    }
-
-    public function write(string $content): self
-    {
-        $this->content .= $content;
-        return $this;
-    }
-
-    public function withHeader(string $name, string $value): self
-    {
-        $this->headers[$name] = $value;
-        return $this;
-    }
-
-    public function status(int $code): self
-    {
-        $this->statusCode = $code;
-        return $this;
-    }
-
-    public function getContent(): string
-    {
-        return $this->content;
-    }
-
-    /**
-     * @return array<string, string>
-     */
-    public function getHeaders(): array
-    {
-        return $this->headers;
-    }
-
-    public function getStatusCode(): int
-    {
-        return $this->statusCode;
-    }
-}
